requirements.txt

pytest==6.0.2

PyYAML==5.3
selenium==3.141.0
